Adding a migration for syncHash, moving genHash to /utilities
